requirements.txt
tensorflow>=2.10
opencv-python
numpy
pillow
streamlit
matplotlib
lxml
